Extract rank to constants
