+++ lib/bothandlers.js
@@ -1,3 +1,4 @@
+import {Rank} from './constants.js';
 import {errorLog} from './logger.js';
 
 /** @typedef {import('./cytubebot.js').CytubeBot} CytubeBot */
@@ -23,7 +24,7 @@ export function addHandlers(bot) {
   bot.socket.on('playlist', (data) => bot.handlePlaylist(data));
 
   bot.socket.on('pm', async (data) => {
-    if (!(await bot.checkPermission(data.username, 2, null))) {
+    if (!(await bot.checkPermission(data.username, Rank.MOD, null))) {
       return;
     }
     bot.handleChatMsg(data, true);
