+++ lib/chatcommands.js
@@ -5,9 +5,8 @@ import humanizeDuration from 'humanize-duration';
 import {promisify} from 'util';
 
 import {callAnagram, callForecast, callGoogleTranslate, callWeather, callWolfram} from './apiclient.js';
-import {RESTART_TIMES_FILE_NAME} from './constants.js';
+import {Rank, RESTART_TIMES_FILE_NAME} from './constants.js';
 import {handle as customHandle} from './custom.js';
-import {Rank} from './cytubebot.js';
 import {auditLog, errorLog, infoLog} from './logger.js';
 import {sendHybridModPermissions} from './permissions.js';
 import {filterMsg, genericUIDLoop, getCurrentUnixTimestamp, getUser, kill, parseBumpData, parseDeleteData, parseForecastData, parseMediaLink, parseUserlimit, sendPmsWithRateLimit, sleep} from './utils.js';
