+++ lib/custom.js
@@ -5,6 +5,7 @@
 
 import random from 'random';
 
+import {Rank} from './constants.js';
 import {sendMessagesWithRateLimit} from './utils.js';
 
 /** @typedef {import('./cytubebot.js').CytubeBot} CytubeBot */
@@ -78,7 +79,7 @@ CUSTOM_HANDLERS.set('spam', async (bot, username, msg) => {
   if (!msg) {
     return;
   }
-  if (!(await bot.checkPermission(username, 2, null))) {
+  if (!(await bot.checkPermission(username, Rank.MOD, null))) {
     return;
   }
