+++ lib/cytubebot.js
@@ -90,15 +90,6 @@ import {Server} from './webserver.js';
  * @property {boolean} paused Whether the video is paused.
  */
 
-export class Rank {
-  static FOUNDER = 5;
-  static OWNER = 4;
-  static ADMIN = 3;
-  static MOD = 2;
-  static REGISTERED = 1;
-  static UNREGISTERED = 0;
-}
-
 export class CytubeBot {
   constructor(config) {
     // init is its own function because constructors can't be async
@@ -271,12 +262,11 @@ export class CytubeBot {
 
     this.talkBot = new Cleverbot;
 
-    this.readPersistentSettings().then((hasPersistentSettings) => {
-      if (!hasPersistentSettings) {
-        this.writePersistentSettings();
-      }
-      this.updatePersistentSettings();
-    });
+    const hasPersistentSettings = await this.readPersistentSettings();
+    if (!hasPersistentSettings) {
+      this.writePersistentSettings();
+    }
+    this.updatePersistentSettings();
 
     if (this.enableWebServer) {
       this.server = new Server(this);
